pyproject.toml
[project]
name = "doscenes-challenge-starter"
version = "0.1.0"
description = "Minimal starter for the doScenes Instructed Driving Challenge"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "scikit-learn>=1.4",
  "sentence-transformers>=2.7",
  "torch>=2.2",
  "tqdm>=4.66",
  "joblib>=1.3",
  "jupyterlab>=4.5.4",
  "matplotlib>=3.10.8",
  "nuscenes-devkit>=1.1.9",
  "pillow>=12.1.1",
  "seaborn>=0.13.2",
]

[tool.uv]
dev-dependencies = []
